downloader: Split the implementation into a module
