+++ src/downloader/mod.rs
@@ -1,0 +1,9 @@
+mod handle;
+mod manager;
+mod types;
+mod worker;
+
+pub use handle::DownloadHandle;
+pub use manager::*;
+pub use types::*;
+pub(self) use worker::download_thread;
